+++ tools/notcurses/utils/ncplayer.sh
@@ -8,6 +8,7 @@ scale() {
 	ncplayer -t 0.5 -s scale ${@}
 }
 
-scale ../../../gnu/gnu.webp
-scale ../../../bpf/cilium/cilium.png
+scale ../../../Documentation/images/cilium.png
+scale ../../../Documentation/images/ebpf.png
+scale ../../../Documentation/images/gnu.webp
 scale ../../../Documentation/images/tux.svg